Рефактор тестов
